requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Computer Vision & ML
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.2

# Tracking
filterpy==1.4.5

# Pose estimation
mediapipe==0.10.7

# Video processing
ffmpeg-python==0.2.0
imageio==2.31.6
imageio-ffmpeg==0.4.9

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Storage
minio==7.2.0
boto3==1.34.0

# Message queue
redis==5.0.1
celery==5.3.4

# WebSocket
websockets==12.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Notifications
requests==2.31.0
python-telegram-bot==20.7
slack-sdk==3.26.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
geopy==2.4.1
